backend/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.14
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
